Use ExternalStorageStrategy for pattern provider pushing logic
